+++ NEWS.md
@@ -1,3 +1,13 @@
+# spectrolab 0.0.14 (2021-08-09)
+
+## major
+* Fixed major bug in read_spectra. The metadata from target and reference where swapped.
+* Now requires R >= 4.0
+
+# Minor
+* Updated citation
+* Updated intro vignette
+
 # spectrolab 0.0.13 (2021-08-05)
 
 ## major